Move AttendancePanel to its own file
