+++ apps/src/code-studio/pd/workshop_dashboard/workshop.jsx
@@ -11,7 +11,6 @@ import React from 'react';
 import {connect} from 'react-redux';
 import moment from 'moment';
 import {Grid, Row, Col, ButtonToolbar, Button} from 'react-bootstrap';
-import {DATE_FORMAT} from './workshopConstants';
 import ConfirmationDialog from '../components/confirmation_dialog';
 import MoveEnrollmentsDialog from './components/move_enrollments_dialog';
 import WorkshopForm from './components/workshop_form';
@@ -21,18 +20,11 @@ import {PermissionPropType, WorkshopAdmin} from './permission';
 import WorkshopPanel from './WorkshopPanel';
 import SignUpPanel from './SignUpPanel';
 import IntroPanel from './IntroPanel';
+import AttendancePanel from './AttendancePanel';
 
 const styles = {
   linkButton: {
     color: 'inherit'
-  },
-  attendanceRow: {
-    padding: '5px 0'
-  },
-  attendanceRowText: {
-    fontSize: '14px',
-    padding: '6px 0',
-    margin: 1
   },
   error: {
     color: 'red',
@@ -593,100 +585,3 @@ const METADATA_FOOTER_STYLE = {
   fontSize: 'smaller',
   fontStyle: 'italic'
 };
-
-class AttendancePanel extends React.Component {
-  static propTypes = {
-    workshopId: PropTypes.string,
-    sessions: PropTypes.array
-  };
-
-  static contextTypes = {
-    router: PropTypes.object.isRequired
-  };
-
-  getSessionAttendanceLink(session) {
-    const url = this.getSessionAttendanceUrl(session);
-    return (
-      <a href={url} target="_blank">
-        {url}
-      </a>
-    );
-  }
-
-  getSessionAttendanceUrl(session) {
-    if (!session.code) {
-      console.warn(`No attendance code found for session ${session.id}`);
-      return null;
-    }
-
-    return `${window.location.protocol}${window.dashboard.CODE_ORG_URL}/pd/${
-      session.code
-    }`;
-  }
-
-  getAttendanceUrl(sessionId) {
-    return `/workshops/${this.props.workshopId}/attendance/${sessionId}`;
-  }
-
-  handleTakeAttendanceClick = event => {
-    event.preventDefault();
-    const sessionId = event.currentTarget.dataset.session_id;
-    this.context.router.push(this.getAttendanceUrl(sessionId));
-  };
-
-  render() {
-    const {sessions} = this.props;
-    return (
-      <WorkshopPanel header="Take Attendance:">
-        <div>
-          <p>
-            There is a unique attendance URL for each day of your workshop. On
-            each day of your workshop, your participants must visit that day's
-            attendance URL to receive professional development credit. The
-            attendance URL(s) will be shown below, 2 days in advance, for your
-            convenience.
-          </p>
-          <Row>
-            <Col md={2}>Date</Col>
-            <Col md={4}>Attendance URL</Col>
-            <Col md={4}>View Daily Roster</Col>
-          </Row>
-          {sessions.map(session => {
-            const date = moment.utc(session.start).format(DATE_FORMAT);
-            return (
-              <Row key={session.id} style={styles.attendanceRow}>
-                <Col md={2}>
-                  <div style={styles.attendanceRowText}>{date}</div>
-                </Col>
-                <Col md={4}>
-                  {session['show_link?'] && (
-                    <div style={styles.attendanceRowText}>
-                      {this.getSessionAttendanceLink(session)}
-                    </div>
-                  )}
-                </Col>
-                <Col md={4}>
-                  <Button
-                    className={
-                      session['show_link?'] && session.attendance_count === 0
-                        ? 'btn-orange'
-                        : null
-                    }
-                    data-session_id={session.id}
-                    href={this.context.router.createHref(
-                      this.getAttendanceUrl(session.id)
-                    )}
-                    onClick={this.handleTakeAttendanceClick}
-                  >
-                    Attendance for&nbsp;
-                    {date}
-                  </Button>
-                </Col>
-              </Row>
-            );
-          })}
-        </div>
-      </WorkshopPanel>
-    );
-  }
-}
